refactor pseudopotentials
move most of the pseudopotential stuff to the pp folder
